requirements.txt
openai
requests
streamlit
pymongo
python-dotenv
fpdf